+++ handlers/Filter.js
@@ -1,3 +1,10 @@
+const blacklist = [
+  "ok",
+  "k",
+  "xani",
+  "ko"
+];
+
 module.exports.filter = (message) => {
   content = message.content.replace(/[\u200B-\u200D\uFEFF]/g, '');
 
@@ -5,7 +12,7 @@ module.exports.filter = (message) => {
   if(currentOkInterval[id]==null){currentOkInterval[id]=1; console.log("new interval entry for channel " + id);}
   if(okSpamLogs[id]==null){okSpamLogs[id]=1; console.log("new entry for channel " + id);}
 
-  if(content.toLowerCase()=="ok"||content.toLowerCase()=="k"){
+  if(blacklist.indexOf(content.toLowerCase())>-1){
     if(okFilter){
       console.log("ok received: " + content);
       okSpamLogs[id]=okSpamLogs[id]+1;currentOkInterval[id] = 0; console.log("ok num increase in channel: " + id +  " new: " + okSpamLogs[id]);
